Kv/macos (#1551)
* fix: enable kotekan compilation on macOS ARM64 (Apple Silicon)

Resolve portability issues preventing kotekan from building on macOS
with Apple Silicon (M3). All changes maintain backward compatibility
with Linux x86_64 builds.

- Rename Metadata.hpp -> NDMetadata.hpp to fix case-insensitive
  filesystem collision with core/metadata.hpp
- Guard x86 intrinsics headers (immintrin.h, emmintrin.h, etc.) with
  architecture checks; all affected files already had runtime fallbacks
- Add sighandler_t typedef for macOS (GNU extension not available)
- Fix size_t vs uint64_t mismatch in CHORDTelescope (different types
  on ARM)
- Guard unconditional omp.h includes with WITH_OMP
- Replace sprintf with snprintf in Config.cpp (deprecated on macOS)
- Add MAX_LOG_MSG_LEN macro to avoid VLA-in-C warning
- Fix missing semicolon in frbPostProcess.cpp non-AVX2 fallback
- Add missing string.h includes in nt_memset.c / nt_memcpy.c fallbacks


* fix: enable LAPACK/Blaze and OpenMP on macOS ARM64

- Add scalar fallbacks for AVX intrinsics in HFBTruncate and
  VisTruncate (replace _mm_malloc/_mm_free with std::malloc/free)
- Guard openblas_set_num_threads() with __APPLE__ in eigenVis.cpp
  (macOS Accelerate BLAS does not provide this API)
- Accept openblas as LAPACKE library in FindLAPACKE.cmake (OpenBLAS
  bundles LAPACKE, no separate liblapacke)
- Link libomp explicitly on macOS (Apple Clang FindOpenMP does not
  set EXE_LINKER_FLAGS)


* fix: enable HDF5, Julia, and Boost tests on macOS ARM64

- Fix size_t vs uint64_t mismatches (different types on ARM macOS):
  hdf5N2Write, fakeGpu, gpuSimulateRFISK
- Replace deprecated std::result_of_t with std::invoke_result_t in
  juliaManager.hpp (deprecated in C++17, errors on newer Apple Clang)
- Guard -latomic link flag with NOT APPLE (built into compiler runtime)
- Remove boost_system from required Boost components (header-only
  since Boost 1.69)
- Add macOS-compatible linker flags for test_Telescope and test_N2Subset
- Add global LAPACKE include path so Blaze can find cblas.h in tests


* fix: static destruction order crash in restServer singleton

Add restServer::is_alive() check so other singletons (datasetManager,
Metrics) can safely skip restServer cleanup if it has already been
destroyed. Fixes a crash on macOS where libc++ enforces strict mutex
validity — Linux/libstdc++ tolerated the use-after-destroy by luck.

The root cause is a classic static destruction order fiasco: restServer
is constructed during datasetManager/Metrics initialization, so it is
destroyed first (reverse order). Their destructors then call
remove_get_callback() on an already-destroyed mutex.


* fix: address PR review feedback for macOS ARM64 port

- Replace __max_log_msg_len variable with MAX_LOG_MSG_LEN define everywhere
- Use __AVX2__ guard instead of __x86_64__ for intrinsics code blocks
- Remove unused mm_malloc.h includes
- Apply lint fix (comment line wrap in stages CMakeLists.txt)
- Add macOS linker flags for new test_NDArrayBufferConfig target


* style: apply auto-lint formatting to test_NDArrayBufferConfig


---------

+++ lib/utils/CHORDTelescope.cpp
@@ -549,7 +549,7 @@ void CHORDTelescope::fill_input_maps(dishInputFields& input) const {
     }
 }
 
-uint64_t CHORDTelescope::get_num_stacks() const {
+size_t CHORDTelescope::get_num_stacks() const {
     FATAL_ERROR("get_num_stacks() has not been implemented in CHORDTelescope yet.");
     return 0;
 }
